requirements.txt
asgiref==3.5.0
certifi==2022.6.15
charset-normalizer==2.0.12
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-phonenumber-field==6.2.0
docopt==0.6.2
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
phonenumbers==8.12.50
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.1
psycopg2-binary==2.9.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.13.4
whitenoise==6.2.0
yarg==0.1.9
